requirements.txt
# Requirements
numpy>=1.16.2
opencv-python>=4.2.0.34
pyzbar>=0.1.8
pytesseract>=0.3.4
pillow>=7.1.2